+++ test/partial_reindex_test.rb
@@ -71,4 +71,14 @@ class PartialReindexTest < Minitest::Test
     end
     assert_match "document missing", error.message
   end
+
+  def test_ignore_missing
+    store [{name: "Hi", color: "Blue"}]
+
+    product = Product.first
+    Product.search_index.remove(product)
+    Searchkick.stub(:ignore_missing, true) do
+      Product.reindex(:search_name)
+    end
+  end
 end