+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 5.2.0 (unreleased)
+## 5.2.0 (2023-02-08)
 
 - Added model name to warning about missing records
 - Fixed unnecessary data loading when reindexing relations with `:async` and `:queue` modes